requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
python-dateutil==2.9.0.post0
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
duckdb==1.1.0
jinja2==3.1.4
pydantic==2.8.2
python-dotenv==1.0.1
reportlab==4.2.2
cryptography==41.0.3
boto3==1.28.52
